requirements.txt
scikit-learn>=1.5,<2
psycopg2-binary>=2.9,<3
qdrant-client>=1.9,<2
chromadb>=0.6,<1
sqlite-vec>=0.1.6,<1
matplotlib>=3.9,<4
seaborn>=0.13,<1
pandas>=2.2,<3
requests>=2.31,<3
numpy>=2.0,<3
psutil>=6.0,<8
